// src/domain/enums/permission.enum.ts
export enum Permission {
    CREATE_USER = 'create_user',
    UPDATE_USER = 'update_user',
    DELETE_USER = 'delete_user',
    READ_USER = 'read_user',

    SET_ROLE = 'set_role',
    SET_PERMISSION = 'set_permission',

    CREATE_DEPARTMENT = 'create_department',
    UPDATE_DEPARTMENT = 'update_department',
    READ_DEPARTMENT = 'read_department',
    DELETE_DEPARTMENT = 'delete_department',

    CREATE_GROUP = 'create_group',
    READ_GROUP = 'read_group',
    UPDATE_GROUP = 'update_group',
    DELETE_GROUP = 'delete_group',

    CREATE_NEWS = 'create_news',
    READ_NEWS = 'read_news',
    UPDATE_NEWS = 'update_news',
    DELETE_NEWS = 'delete_news',
}